LOCAL-1: Refactored and added TODOs
