+++ classes/your_team.py
@@ -1,0 +1,15 @@
+# TODO: import the required parent class SportsTeam
+# ================================ YOUR CODE HERE ============================== #
+
+
+# TODO: Create the class YourTeam with all required parent class init parameters, as well as the
+#  team_name and best_player attributes
+# ================================ YOUR CODE HERE ============================== #
+
+# TODO: Create a 'get_best_player' method that returns the string 'The best player on the {team name} is {best player}'
+# ================================ YOUR CODE HERE ============================== #
+
+
+# TODO: Create a 'get_national_anthem' method that returns the string
+#  'The first 5 words of the anthem from {country} are {anthem 5 words}
+# ================================ YOUR CODE HERE ============================== #
